+++ db/migrations/201904261359_add_extended_columns_to_labels.rb
@@ -1,0 +1,6 @@
+Sequel.migration do
+  change do
+    extension :pg_json
+    add_column :labels, :extended_columns, :jsonb
+  end
+end